Move `MagneticGeometry` to new file
