+++ src/fusiondls/__init__.py
@@ -8,7 +8,8 @@ except PackageNotFoundError:
     __version__ = get_version(root="..", relative_to=__file__)
 
 from .AnalyticCoolingCurves import LfuncN
-from .DLScommonTools import MagneticGeometry, file_read, file_write, make_arrays
+from .DLScommonTools import file_read, file_write, make_arrays
+from .geometry import MagneticGeometry
 from .LRBv21 import run_dls
 
 __all__ = [
